Convert ResumableUploader to a plugin
